Bug 1135768 - Review comment: remove title blur and corner radius popping.
The title blur popped terribly depending on the colors at the top of the
page in the tab.  I think this is due to the blur picking colors first
from the static image, and then from the web view itself.  At some point
we can re-instate this effect; but for now, it's better to be opaque and
clean than blurred, fancy, and dirty.