Bug 1195204: Fix location bar elements on iOS 9
Strangely, on iOS 9 setting `hidden` does not suffice in some situations
to hide a view when that view is returned from
`accessibilityElements`. So let us return only non-hidden views from
`accessibilityElements` property. This does fix the bug.

Note that such a behavior is not present when having a minimal example
of view controller with bunch of buttons and labels as direct subviews
of the VC's `view`, some of them `hidden` and some of them not, all
returned by `accessibilityElements` - there, still only the non-`hidden`
views are visible to VoiceOver. So it is interesting that the location
bar exhibits this problem.

While there, let's be safe and do post an `AXLayoutChanged` notification
when the `hidden` status of one of the views changes. Maybe in iOS
10 (or rather iOS X ?), things would stop working if we did not :-),
just like they did in this case in iOS 9.