Bug 1283143 - disable privacy mode toggle while adding a new tab (https://github.com/mozilla-mobile/firefox-ios/issues/2115) r=fpatel
A crash was being caused when toggling privacy mode while opening a new tab. This caused the new tab to be opened in the wrong collection as the mode switch happened before the batch updates completed, causing a crash (assertion failure in CollectionViewData).
Therefore I am disabling the privacy toggle button for the duration of the add tab action, preventing this situation from ever occurring.