Bug 1323373 - Don't upload tens of thousands of history items at once. r=sleroux
We can't simply pull every row from the DB, turn it into a record, and
dump it on the upload batcher -- we rapidly end up with hundreds of
thousands of allocated objects, as we create JSON blobs for each visit,
and a Record and a Payload, as well as retaining all of the original
data pulled from the DB.

This commit has us chunk the data from the DB, processing only 1000
records at a time. Each batch contains no more than 1000 records,
with timestamps chained through. Each POST will still include no more
than 100 records, thanks to existing limitations.

Uploading 80,000 records will still take 800 POSTs -- we just won't
die with a memory error in the attempt.