Bug 1355214 - Add stage servers debug option. r=sleroux (#2617)
* Bug 1355214 - Pre: Don't default ChinaSyncServiceSetting to on.

This is just hygiene -- the control will never be shown unless
`isChinaEdition` is true -- but if it _is_ shown, it will end up
enabling the preference as a side effect.  That's not right.

* Bug 1355214 - Add stage servers debug option. r=sleroux

I initially extracted a subclass so that
{China,Stage}SyncService{Debug}Setting would share code, but then I
wanted to update the status text to reflect the fxa-auth-server
endpoint URL, and the code diverged enough to not be worth it.  So,
duplication rather than lots of overrides.