Bug 1395174: iOS 11 table cell text color can conflict with NSAttributedString color (#3122)
In cellForRowAt we are setting disable state by both setting the cell.textLabel.color
and the color attribute on NSAttributedString. On iOS 11 we see that the latter is trumping
during first creation, and then on cell reuse the former is applied.

The fix is easy. But to protect against future misuse, I am changing our utility function for getting
table row text to force the developer to decide if you want the enabled or disabled color.
If they then go and combine this with cell.textLabel.color, well, good luck to them.