Bug 1434695 - Load cached clients when showing the client picker in Send Tab. (https://github.com/mozilla-mobile/firefox-ios/issues/3795) r=justindarc
Previously we would kick off a refresh spinner, do a full sync of clients, then fetch them from the DB. Only then would we display clients to choose.

Most users' client lists will almost never change — certainly not multiple times a day. The clients synchronizer will debounce a little, but it still does an extraordinary amount of work to even get to that point.

With this PR, we first fetch the client list from the DB. If there are no clients, we kick off a regular refresh with spinner. If there are clients, we silently run a client sync (no spinner) and update the displayed list if it changes.

Pull-to-refresh is unaffected.