Bug 1487165 - Tab store async off-main writes, and scoped lock for reads (#4181)
* Remove tabManagerWillStoreTabs from appDelegate
This should be in the TabManager. Just requires having a Profile
reference in the TabManager.

* Remove isRestore flag determining if delegates called.
- makes this a responsibility of the delegate to decide if the want to take
action. The originating code is not the decision-maker as to whether the
delegates should take action.
- first step in getting rid of the isRestore state flag. A scoped
lock should be used, not a toggled flag.

* Rename toTab to toRemoteTab, which is clearer.
And make it clearer (and safer) early in the call stack that this is not for private
tabs. They will never become remote tabs.

* Tab store is now async on write, and locked while reading
- writing tabs returns a deferred, if needed one can call code on
completion (returning a deferred here is primarily useful to indicate the async
nature of the call). TODO: this is called too often in existing code, but is outside of
the scope of this PR.
- restoring tabs is now an internally managed scoped lock, no more toggling+managing restoring state flags, while in the
scoped lock, the tabs are restoring.  Restoring tabs is unchanged in
terms of main-thread-blocking, this makes sense to block while this
operation is called.

Additional cleanup:
- restoring tabs clearly differentiates between restoring tabs at
startup, versus restoring tabs for undo
- startup tab restoration would read the list of tabs from disk everytime the tab
store was queried for the boolean check to see if tabs were available
for restoration, a lazy var for `archivedStartupTabs` avoids any
problems by initializing this once, and then clearing it when no longer
needed

* Fix test cases to match new API

* change dispatch time to new syntax