Fix https://github.com/mozilla-mobile/firefox-ios/issues/5795 - Error page not shown on reload (#5800)
The old reload logic was to use a timestamp to determine if the error
page was being loaded or reloaded. The idea was that if errorpage load
was called with a timestamped URL <2s, load the errorpage normally; if
errorpage load was called with a timestamped URL >2s old then reload the
original page. It allowed for one codepath to handle the reload logic,
but it is very flaky.
Instead just have the reload buttons have special behaviour to reload
the original URL.