Update event recording for Glean v39. (https://github.com/mozilla-mobile/firefox-ios/issues/8919)
Glean v39 brings a new API for recording typed event extras.
The firefox-ios code uses exactly one typed event extra, for
recording the slide number on which the onboarding flow was
dismissed. This updates the recording code for that metric
to use the new APIs.

We'll need this when updating to the next release of appservices,
which brings with it the new Glean. Unfortunately I couldn't find
a way to change the code to be compatible with the previous version
of Glean, so we won't be able to land this in the meantime.

(Also, as far as I can the previous code was not recording
this property correctly anyway - it was looking for the slide
number as the "EventValue" argument but it's actually being sent
in via the "extras" argument. I've adjusted the code to pull it
from the right place and confirmed via debug logging that it's
recording the expected value.)